[Lens] cleanup divide mock file to smaller pieces
